refactor: major minor level code smell 일부 제거
