+++ backend/goodday/build.gradle
@@ -78,7 +78,6 @@ jacoco{
 	toolVersion = "0.8.7"
 }
 
-
 jacocoTestReport{
 	reports {
 		html.enabled true
