+++ backend/goodday/src/main/java/seeuthere/goodday/board/controller/BoardController.java
@@ -15,7 +15,7 @@ import seeuthere.goodday.auth.domain.EnableAuth;
 import seeuthere.goodday.board.domain.Board;
 import seeuthere.goodday.board.domain.Comment;
 import seeuthere.goodday.board.dto.request.BoardRequest;
-import seeuthere.goodday.board.dto.request.BoardResponse;
+import seeuthere.goodday.board.dto.response.BoardResponse;
 import seeuthere.goodday.board.dto.request.CommentRequest;
 import seeuthere.goodday.board.service.BoardService;
 import seeuthere.goodday.member.domain.Admin;
