+++ backend/goodday/src/main/java/seeuthere/goodday/board/dto/response/BoardResponse.java
@@ -1,4 +1,4 @@
-package seeuthere.goodday.board.dto.request;
+package seeuthere.goodday.board.dto.response;
 
 import java.time.LocalDateTime;
 import seeuthere.goodday.board.domain.Board;
@@ -6,26 +6,16 @@ import seeuthere.goodday.board.domain.BoardLabel;
 
 public class BoardResponse {
 
-    private final Long id;
-    private final String title;
-    private final String content;
-    private final LocalDateTime createTime;
-    private final LocalDateTime updateTime;
-    private final BoardLabel label;
-    private final String memberId;
-    private final CommentResponse commentResponse;
+    private Long id;
+    private String title;
+    private String content;
+    private LocalDateTime createTime;
+    private LocalDateTime updateTime;
+    private BoardLabel label;
+    private String memberId;
+    private CommentResponse commentResponse;
 
-    public BoardResponse(Long id, String title, String content, LocalDateTime createTime,
-        LocalDateTime updateTime, BoardLabel label, String memberId,
-        CommentResponse commentResponse) {
-        this.id = id;
-        this.title = title;
-        this.content = content;
-        this.createTime = createTime;
-        this.updateTime = updateTime;
-        this.label = label;
-        this.memberId = memberId;
-        this.commentResponse = commentResponse;
+    public BoardResponse() {
     }
 
     public BoardResponse(Board board) {
